FEAT: api test cases added

+++ apps/users/serializers.py
@@ -7,7 +7,7 @@ class EmployeeStatisticsSerializer(serializers.ModelSerializer):
     full_name = serializers.CharField(source='get_full_name')
     client_count = serializers.IntegerField(default=0)
     total_product_count = serializers.IntegerField(default=0)
-    total_order_sum = serializers.IntegerField(default=0)
+    total_order_sum = serializers.DecimalField(max_digits=20, decimal_places=2, default=0)
 
     class Meta:
         model = models.Employee
@@ -23,7 +23,7 @@ class EmployeeStatisticsSerializer(serializers.ModelSerializer):
 class ClientStatisticsSerializer(serializers.ModelSerializer):
     full_name = serializers.CharField(source='get_full_name')
     total_bought_product_count = serializers.IntegerField(default=0)
-    total_order_sum = serializers.IntegerField(default=0)
+    total_order_sum = serializers.DecimalField(max_digits=20, decimal_places=2, default=0)
 
     class Meta:
         model = models.Client
